requirements.txt
# ===============================
# Core NLP / Deep Learning (CPU)
# ===============================

torch==2.1.2
transformers==4.37.2
datasets==2.16.1
accelerate==0.27.2

# ===============================
# MLOps / Tracking
# ===============================

mlflow==2.10.0

# ===============================
# Data / Scientific stack
# ===============================

numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.0
pyarrow==14.0.2

# ===============================
# API / Serving
# ===============================

fastapi==0.110.0
uvicorn[standard]==0.27.1

# ===============================
# Monitoring / Logs (Azure)
# ===============================

opencensus==0.11.4


# ===============================
# Utils
# ===============================

tqdm==4.66.1
typing-extensions==4.9.0
sentencepiece==0.2.0
safetensors==0.4.5

# Test dependencies
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
opencensus-ext-azure==1.1.9
opencensus-ext-requests==0.8.0

